fallback to inline loops
