+++ libnd4j/include/loops/cpu/reduce/reduce_same.cpp
@@ -197,7 +197,10 @@ namespace functions {
                     tadOnlyShapeInfo = tadPack.primaryShapeInfo();
                     tadOffsets = tadPack.primaryOffsets();
                 }
-                nd4j::Loops::loopTadXZ<X, X, X, OpType>(x, xShapeInfo, z, zShapeInfo,  tadOnlyShapeInfo, tadOffsets, dimension, dimensionLength, extraParams);
+
+                nd4j::Loops::loopTadXZ<X, X, X, OpType>(const_cast<const X*>(x), const_cast<const Nd4jLong *>(xShapeInfo), z,
+                                                        const_cast<const Nd4jLong *>(zShapeInfo), const_cast<const Nd4jLong *>(tadOnlyShapeInfo),
+                                                        const_cast<const Nd4jLong *>(tadOffsets), const_cast<const int *>(dimension), dimensionLength, extraParams);
             }
